# apiDados/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api Hypatiamat",
        "description": "Aqui encontra-se a documentação de toda a api de dados do Hypatiamat, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3050",
    "basePath": "/",
...
    "paths": {
        "/aplicacoes/login":{
          "post":{
            "tags":[
                "Professores",
                "Alunos"
            ],
            "summary": "Autenticar um professor ou aluno.",
            "description": "Autenticação de um utilizador.",
            "parameters": [
                {
                  "name": "user",
                  "in": "formData",
                  "description": "Código/Username de um professor/aluno",
                  "required": true,
                  "type": "string"
                },
                {
                  "name": "password",
                  "in": "formData",
                  "description": "Password do professor/aluno",
                  "required": true,
                  "type": "string",
                  "format": "password"
                }
              ],
            "responses": {
                "200": {
                    "description": "Retorna o token de acesso e o tipo de acesso caso as credenciais sejam válidas",
                    "schema": {
                        "type": "object",
                            "properties": {
                                "authentication": {
                                    "type": "boolean"
                                },
                                "token":{
                                  "type": "string"
                                },
                                "type":{
                                  "type": "integer"
                                }
                            }
                        }
                }
            }
          }
        },
        "/aplicacoes/professores": {
            "get": {
                "tags":[
                    "Professores"
                ],
                "summary": "Todos os professores registados.",
                "description": "Retorna uma lista com toda informação à cerca de todos os professores.",
                "responses": {
                    "200": {
                        "description": "Lista de Professores",
                        "schema": {
                          "type":"array",
                          "items":{
                            "$ref": "#/definitions/Professor"
                          }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "post":{
                "tags":[
                    "Professores"
                ],
                "summary": "Inserir um novo professor.",
                "description": "Adiciona um novo curso à base de dados.",
                "parameters": [
                    {
                      "name": "codigo",
                      "in": "formData",
                      "description": "Código/Username de um professor",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do professor",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "escola",
                      "in": "formData",
                      "description": "Identificador da escola do professor",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "email",
                      "in": "formData",
                      "description": "Email do professor",
                      "required": true,
                      "type": "string"
                    },{
                      "name": "password",
                      "in": "formData",
                      "description": "Password do professor",
                      "required": true,
                      "type": "string",
                      "format": "password"
                    },{
                      "name": "confirmacao",
                      "in": "formData",
                      "description": "0 se o professor não está confirmado, 1 caso contrário",
                      "required": true,
                      "type": "integer"
                    },{
                      
                      "name": "premium",
                      "in": "formData",
                      "description": "Nível de acesso",
                      "required": true,
                      "type": "integer"
                    },{
                      "name": "socionum",
                      "in": "formData",
                      "description": "Número de sócio do professor",
                      "required": true,
                      "type": "integer"
                    },{
                      "name": "validade",
                      "in": "formData",
                      "description": "Data em que a sua conta expira",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "projeto",
                      "in": "formData",
                      "description": "Projeto",
                      "required": true,
                      "type": "integer"
                    }

                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do professor criado",
                        "schema": {
                            "type": "string",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
          "/aplicacoes/alunos": {
            "get": {
                "tags":[
                    "Alunos"
                ],
                "summary": "Todos os alunos registados.",
                "description": "Retorna uma lista com a informação à cerca de todos os alunos.",
                "responses": {
                    "200": {
                        "description": "Lista de Alunos",
                        "schema": {
                          "type":"array",
                          "items":{
                            "$ref": "#/definitions/Aluno"
                          }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "post":{
                "tags":[
                    "Alunos"
                ],
                "summary": "Inserir um novo aluno.",
                "description": "Adiciona um novo aluno à base de dados.",
                "parameters": [
                    {
                      "name": "user",
                      "in": "formData",
                      "description": "Código/Username do aluno",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "numero",
                      "in": "formData",
                      "description": "Número do aluno",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do aluno",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "datanascimento",
                      "in": "formData",
                      "description": "Data de nascimento do aluno",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "escola",
                      "in": "formData",
                      "description": "Identificador da escola do aluno",
                      "required": true,
                      "type": "integer"
                    },
                    {
                      "name": "turma",
                      "in": "formData",
                      "description": "Turma do aluno",
                      "required": true,
                      "type": "string"
                    },
                    {
                      "name": "email",
                      "in": "formData",
                      "description": "Email do aluno",
                      "required": true,
                      "type": "string"
                    },{
                      "name": "password",
                      "in": "formData",
                      "description": "Password do aluno",
                      "required": true,
                      "type": "string",
                      "format": "password"
                    },{
                      "name": "codprofessor",
                      "in": "formData",
                      "description": "Código de acesso fornecido pelo professor",
                      "required": true,
                      "type": "string"
                    },{
                      
                      "name": "pais",
                      "in": "formData",
                      "description": "País do aluno",
                      "required": true,
                      "type": "string"
                    },{
                      "name": "confirmacao",
                      "in": "formData",
                      "description": "1 caso esteja confirmado, 0 caso contrário",
                      "required": true,
                      "type": "integer"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do aluno criado",
                        "schema": {
                            "type": "string",
                            "items": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/aplicacoes/escolas": {
          "get": {
              "tags":[
                  "Escolas"
              ],
              "summary": "Todos as escolas registados.",
              "description": "Retorna uma lista com a informação à cerca das escolas.",
              "responses": {
                  "200": {
                      "description": "Lista de Escolas",
                      "schema": {
                        "type":"array",
                        "items":{
                          "$ref": "#/definitions/Escola"
                        }
                      }
                  },
                  "401":{
                    "description": "Falta-lhe o token de acesso ou o token não é válido."
                  }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
          },
          "post":{
              "tags":[
                  "Escolas"
              ],
              "summary": "Inserir uma nova escola.",
              "description": "Adiciona uma nova escola à base de dados.",
              "parameters": [
                  {
                    "name": "nome",
                    "in": "formData",
                    "description": "Nome da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "localidade",
                    "in": "formData",
                    "description": "Localidade da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "distrito",
                    "in": "formData",
                    "description": "Distrito da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "pais",
                    "in": "formData",
                    "description": "País da escola",
                    "required": true,
                    "type": "string"
                  },
                  {
                    "name": "cod",
                    "in": "formData",
                    "description": "Código da escola",
                    "required": true,
                    "type": "string"
                  }
                ],
              "responses": {
                  "200": {
                      "description": "Retorna o id da escola criada",
                      "schema": {
                          "type": "string",
                          "items": {
                              "properties": {
                                  "id": {
                                      "type": "string"
                                  }
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/aplicacoes/turmas/":{
        "get": {
          "tags": [
            "Turmas"
          ],
          "summary": "Devolve todas as turmas do sistema.",
          "description": "Retorna todas as turmas registadas.",
          "operationId": "getTurmas",
          "responses": {
            "200": {
                "description": "Retorna a lista de turmas.",
                "schema": {
                    "type": "array",
                    "items":{
                      "$ref": "#/definitions/Turma"
                    }   
                    }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        },
        "post":{
          "tags": [
            "Turmas"
          ],
          "summary": "Inserir uma nova turma.",
          "description": "Criar uma nova turma.",
          "operationId": "insertTurma",
          "parameters": [
            {
              "name": "turma",
              "in": "formData",
              "description": "Nome da turma.",
              "required": true,
              "type": "string"
            },
            {
              "name": "idprofessor",
              "in": "formData",
              "description": "Código/Username do professor da turma.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "A turma foi criada e é devolvido o seu identificador.",
              "schema": {
                "type": "string",
                "items": {
                    "properties": {
                        "id": {
                            "type": "string"
                        }
                    }
                }
              }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      },
...
    },
    "securityDefinitions":{
      "api_key":{
        "type": "apiKey",
        "in": "query",
        "name": "token"
      }
    },
    "security": [{
       "api_key": []
    }],
    "definitions": {
        "Aluno": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "user": {
                "type": "string"
              },
              "numero": {
                "type": "integer"
              },
              "nome": {
                "type": "string"
              },
              "datanascimento": {
                "type": "string"
              },
              "escola": {
                "type": "string"
              },
              "turma":{
                "type": "string"
              },
              "email":{
                "type":"string"
              },
              "confirmacao":{
                "type":"integer"
              }
            }
        },
        "Professor":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "codigo": {
              "type": "string"
            },
            "nome": {
              "type": "string"
            },
            "escola": {
              "type": "string"
            },
            "email":{
              "type":"string"
            },
            "confirmacao":{
              "type":"integer"
            },
            "premium":{
              "type":"integer"
            },
            "validade":{
              "type":"string"
            },
            "socionum":{
              "type":"integer"
            },
            "projeto":{
              "type":"integer"
            }
          }    
        },
        "Turma":{
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "idprofessor": {
              "type": "string"
            },
            "turma":{
              "type":"string"
            }
          }
        },
        "Escola":{
          "type": "object",
          "properties": {
            "id":{
              "type": "integer"
            },
            "nome":{
              "type":"string"
            },
            "localidade":{
              "type":"string"
            },
            "distrito":{
              "type":"string"
            },
            "pais":{
              "type":"string"
            },
            "cod":{
              "type":"string"
            }
          }
        }
    }   
}